Changes to move event handling registration out of TTTController and into GUIView
TTTController was the wrong place for registering the event handlers because
it moved JavaFX libraries into the TTTController which shouldn't know about
JavaFx.
The reason I put this event handler reg in TTTController was because I was
trying to avoid passing the controller instance to the event handlers, however
it looks like it's a necessary evil as even using Observer patern I would need
to pass the Observer to the Observable.
