+++ GUI/src/javafxgui/ClickEventHandler.java
@@ -1,5 +1,5 @@
 package javafxgui;
 
 public interface ClickEventHandler {
-    void action(String id);
+    void action(String displayPositionId);
 }
