+++ GUI/src/javafxgui/EventRegister.java
@@ -6,27 +6,27 @@ import javafx.scene.control.Button;
 import javafx.scene.layout.GridPane;
 
 public class EventRegister {
-    public void registerClickableElementWithHandler(ClickableElement clickableElement, ClickEventHandler eventHandler) {
-        clickableElement.setOnAction(eventHandler);
+    public void registerAllClickableElementsWithHandler(Scene scene, GUIView guiView) {
+        registerBoard(extractGameBoard(scene), guiView);
+        registerReplayButtonWithHandler(extractReplayButton(scene), guiView);
     }
 
-    public void registerAllBoardButtonsWithHandler(GridPane board, Controller controller) {
+    public void registerAllBoardButtonsWithHandler(GridPane board, GUIView guiView) {
         for (Node button : board.getChildren()) {
-            registerClickableElementWithHandler(new JavaFXButton((Button) button), new NewPlayerMoveEventHandler(controller));
+            registerClickableElementWithHandler(new JavaFXButton((Button) button), new NewPlayerMoveEventHandler(guiView));
         }
     }
 
-    public void registerReplayButtonWithHandler(Button replayButton, Controller controller) {
-        registerClickableElementWithHandler(new JavaFXButton(replayButton), new NewGameEventHandler(controller));
+    private void registerBoard(GridPane gameBoard, GUIView guiView) {
+        registerAllBoardButtonsWithHandler(gameBoard, guiView);
     }
 
-    public void registerAllClickableElementsWithHandler(Scene scene, Controller controller) {
-        registerBoard(extractGameBoard(scene), controller);
-        registerReplayButtonWithHandler(extractReplayButton(scene), controller);
+    private void registerReplayButtonWithHandler(Button replayButton, GUIView guiView) {
+        registerClickableElementWithHandler(new JavaFXButton(replayButton), new NewGameEventHandler(guiView));
     }
 
-    private void registerBoard(GridPane gameBoard, Controller controller) {
-        registerAllBoardButtonsWithHandler(gameBoard, controller);
+    private void registerClickableElementWithHandler(ClickableElement clickableElement, ClickEventHandler eventHandler) {
+        clickableElement.setOnAction(eventHandler);
     }
 
     private Button extractReplayButton(Scene scene) {
