+++ GUI/src/javafxgui/TTTController.java
@@ -1,7 +1,5 @@
 package javafxgui;
 
-import javafx.scene.Scene;
-import javafx.scene.layout.GridPane;
 import jttt.Core.Board.Board;
 import jttt.Core.Game;
 import jttt.Core.Players.PlayerFactory;
@@ -9,43 +7,40 @@ import jttt.Core.Players.PlayerFactory;
 public class TTTController implements Controller {
     private final int DEFAULT_BOARD_DIMENSION = 3;
     private final int HVH_GAMETYPE = 1;
-    private EventRegister eventRegister;
-    private GUIDisplay gameView;
+    private GUIView guiView;
     private Game game;
 
-    public TTTController(GUIDisplay gameView, EventRegister eventRegister, Game game) {
-        this.gameView = gameView;
-        this.eventRegister = eventRegister;
-        this.game = game;;
+    public TTTController(GUIView guiView, Game game) {
+        this.guiView = guiView;
+        guiView.setController(this);
+        this.game = game;
     }
 
-    public Scene displayGUI() {
-        Scene scene = gameView.displayGUI(game.getBoard());
-        registerEventHandlers(scene);
-        return scene;
+    public void displayGUI() {
+        guiView.displayGUI(game.getBoard());
     }
 
-    public GridPane displayBoard() {
-        return gameView.displayBoard(game.getBoard());
+    public void displayBoard() {
+        guiView.displayBoard(game.getBoard());
     }
 
     public void displayResult() {
         if (foundWinOrDraw()){
-            gameView.disableBoard(game.getBoard());
-            gameView.displayResult(game.findWinner());
+            guiView.disableBoard(game.getBoard());
+            guiView.displayResult(game.findWinner());
         }
     }
 
     public void playMoveAtPosition(String id) {
         playMoveOnGameBoard(id);
-        registerBoardEventHandlers(displayBoard());
+        displayBoard();
         displayResult();
         displayPlayAgain();
     }
 
     public void createNewGame() {
         clearGameBoard();
-        registerEventHandlers(gameView.displayGUI(game.getBoard()));
+        guiView.displayGUI(game.getBoard());
     }
 
     public boolean foundWinOrDraw() {
@@ -54,16 +49,8 @@ public class TTTController implements Controller {
 
     public void displayPlayAgain() {
         if (foundWinOrDraw()){
-            gameView.makePlayAgainVisible();
+            guiView.makePlayAgainVisible();
         }
-    }
-
-    private void registerEventHandlers(Scene scene) {
-        eventRegister.registerAllClickableElementsWithHandler(scene, this);
-    }
-
-    private void registerBoardEventHandlers(GridPane board) {
-        eventRegister.registerAllBoardButtonsWithHandler(board, this);
     }
 
     private void playMoveOnGameBoard(String id) {
@@ -72,6 +59,6 @@ public class TTTController implements Controller {
 
     private void clearGameBoard() {
         game = new Game(new Board(DEFAULT_BOARD_DIMENSION), HVH_GAMETYPE, new PlayerFactory());
-        gameView = new GUIDisplay(gameView.getScene(), new BoardDisplay());
+//        guiView = new GUIView(guiView.getScene(), new BoardDisplay(), new EventRegister());
     }
 }
