+++ GUI/test/javafxgui/EventHandlerTest.java
@@ -2,9 +2,6 @@ package javafxgui;
 
 import javafx.scene.Scene;
 import javafx.scene.layout.StackPane;
-import jttt.Core.Board.Board;
-import jttt.Core.Game;
-import jttt.Core.Players.PlayerFactory;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -12,27 +9,24 @@ import static org.junit.Assert.assertEquals;
 
 public class EventHandlerTest {
 
-    private TTTControllerSpy controllerSpy;
+    private GUIViewSpy guiViewSpy;
 
     @Before
     public void setUp() {
-        controllerSpy = new TTTControllerSpy(
-                new GUIDisplay(new Scene(new StackPane(),700, 600), new BoardDisplay()),
-                new EventRegister(),
-                new Game(new Board(3), 1, new PlayerFactory()));
+        guiViewSpy = new GUIViewSpy(new Scene(new StackPane(), 700, 600), new BoardDisplay(), new EventRegister());
     }
 
     @Test
     public void handleNewMoveEvent() {
-        NewPlayerMoveEventHandler startHandler = new NewPlayerMoveEventHandler(controllerSpy);
+        NewPlayerMoveEventHandler startHandler = new NewPlayerMoveEventHandler(guiViewSpy);
         startHandler.action("1");
-        assertEquals(true, controllerSpy.hasReDisplayBoardBeenCalled());
+        assertEquals(true, guiViewSpy.hasBoardButtonBeenClicked());
     }
 
     @Test
     public void handleRestartGameEvent() {
-        NewGameEventHandler newGameHandler = new NewGameEventHandler(controllerSpy);
+        NewGameEventHandler newGameHandler = new NewGameEventHandler(guiViewSpy);
         newGameHandler.action("");
-        assertEquals(true, controllerSpy.hasReplayGameBeenSelected());
+        assertEquals(true, guiViewSpy.hasReplayGameBeenSelected());
     }
 }
